Pull out contained columns as a base class
